// frontend/src/shared/types/api.ts
export interface components {
    schemas: {
// ...
        MessageRead: {
            /** Artifacts */
            artifacts: {
                [key: string]: unknown;
            }[];
            /** Citations */
            citations: {
                [key: string]: unknown;
            }[];
            /** Content */
            content: string;
            /**
             * Conversation Id
             * Format: uuid
             */
            conversation_id: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Role */
            role: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}